Cleanup composite provider
