+++ scanner/scanner/providers/provider.py
@@ -35,6 +35,42 @@ class Provider(ABC):
 	async def get_serie(self, external_id: dict[str, str]) -> Serie | None:
 		raise NotImplementedError
 
+	async def find_movie(
+		self, title: str, year: int | None, external_id: dict[str, str]
+	) -> Movie:
+		ret = await self.get_movie(external_id)
+		if ret is not None:
+			return ret
+		search = await self.search_movies(title, year, language=[])
+		if not any(search):
+			raise ProviderError(
+				f"Couldn't find a movie with title {title}. (year: {year}"
+			)
+		ret = await self.get_movie(
+			{k: v.data_id for k, v in search[0].external_id.items()}
+		)
+		if not ret:
+			raise ValueError()
+		return ret
+
+	async def find_serie(
+		self, title: str, year: int | None, external_id: dict[str, str]
+	) -> Serie:
+		ret = await self.get_serie(external_id)
+		if ret is not None:
+			return ret
+		search = await self.search_series(title, year, language=[])
+		if not any(search):
+			raise ProviderError(
+				f"Couldn't find a serie with title {title}. (year: {year}"
+			)
+		ret = await self.get_serie(
+			{k: v.data_id for k, v in search[0].external_id.items()}
+		)
+		if not ret:
+			raise ValueError()
+		return ret
+
 
 class ProviderError(RuntimeError):
 	def __init__(self, *args: object) -> None:
